requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
pytesseract
python-dotenv
httpx
pytest
pytest-asyncio
gunicorn
